Add victory panel and unify game status HUD controller
Introduces a new victory panel to the GameHUD with corresponding styles and UI elements. Refactors GameOverHUDController to GameStatusHUDController to handle both game over and victory states, updating references and logic accordingly. Updates material sun direction values and scene references to match the new controller.
