--- Red Strike/Assets/UISystem/GameOverHUDController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 508797d78c25d0a41a624e12108fe2c7
