+++ Red Strike/Assets/UISystem/GameStatusHUDController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 14c1e9364681d8a4dbb53c6fbb5dbf6f
